Moves Setup_Hooks to its own compilation unit.
