+++ src/CMakeLists.txt
@@ -28,7 +28,7 @@ set(GAMEENGINE_INCLUDES
     w3d/renderer
 )
 
-set(HOOKER_SRC hooker/dllmain.cpp hooker/hooker.cpp)
+set(HOOKER_SRC hooker/dllmain.cpp hooker/hooker.cpp hooker/setuphooks.cpp)
 
 set(GAMEENGINE_SRC
     game/copyprotect.cpp
